+++ CHANGELOG.md
@@ -78,3 +78,4 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - fix/1201234903595047 - Fix toArray from CollectionInterfaceTrait
 - fix/1201352833985547 - When iterating on a collection and modifying the collection infinite recursion will happen
 - fix/1201438240911772 - Add additional checks to sorting traits
+- fix/1201487027413419 - Allow mixed return types for factory interfaces